requirements.txt
aioredis==2.0.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==5.0.1
bcrypt==4.1.2
CacheControl==0.14.2
cachetools==5.5.0
certifi==2025.1.31
cffi==1.16.0
charset-normalizer==3.4.1
click==8.1.7
cryptography==42.0.2
et_xmlfile==2.0.0
fastapi==0.109.2
firebase-admin==6.6.0
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.20.0
google-cloud-storage==3.0.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
greenlet==3.0.3
grpcio==1.70.0
grpcio-status==1.70.0
gunicorn==21.2.0
h11==0.14.0
httplib2==0.22.0
idna==3.6
msgpack==1.1.0
openpyxl==3.1.5
packaging==23.2
pillow==11.1.0
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
PyJWT==2.10.1
PyMySQL==1.1.0
PyNaCl==1.5.0
pyparsing==3.2.1
python-dotenv==1.0.1
python-multipart==0.0.20
redis==5.2.1
requests==2.32.3
rsa==4.9
sniffio==1.3.0
SQLAlchemy==2.0.27
starlette==0.36.3
stripe==13.0.1
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.27.1
whitenoise==6.8.2
